refactor(metrics): move vsock metrics out of logger
Moving vsock metrics out of logger helps decouple
vsock metrics from logger and be consistent with block
and net which too have moved their metrics out of logger.
