+++ src/vmm/src/devices/virtio/vsock/mod.rs
@@ -14,6 +14,7 @@
 mod csm;
 mod device;
 mod event_handler;
+pub mod metrics;
 mod packet;
 pub mod persist;
 pub mod test_utils;
